pyproject.toml
[project]
name = "echo-lifesim"
version = "0.1.0"
description = "ECHO-LifeSim MVP: Mirror-mode persona simulation with episodic memory, needs, and A/B action suggestions."
authors = [ { name="You" } ]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.12.3",
    "orjson>=3.10.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy", "pytest-cov"]
gui = ["streamlit>=1.36.0"]

[project.scripts]
echo-sim = "echo_lifesim.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true

